Migrate _utils/Item to Styled Components (#313)
