+++ package.json
@@ -69,6 +69,7 @@
     "gh-pages": "2.0.1",
     "husky": "1.3.1",
     "jest": "24.8.0",
+    "jest-styled-components": "6.3.3",
     "jsdom": "14.0.0",
     "jsdom-global": "3.0.2",
     "prettier": "1.16.4",
@@ -88,7 +89,6 @@
     "country-telephone-data": "0.6.0",
     "exenv": "1.2.2",
     "focus-trap": "5.0.0",
-    "jest-styled-components": "6.3.3",
     "lodash.debounce": "4.0.8",
     "lodash.isempty": "4.4.0",
     "lodash.isequal": "4.5.0",
